frontend-app-api: extract extension config reading into separate module
